Add ParentCat to move cln/topic cats for all langs
Create a new class ParentCat that can move all of the
language-specific categories within a catlangname or topic category.
For example I used it to move "Category:Racist names for places" and
its subcategories such as "Category:en:Racist names for places" to
"Category:Derogatory names for places" (and "Category:en:Derogatory
names for places").