Bug: Save whole page, not just specified language
Previously, if a language was specified, for the sake of efficiency
only the section for the specified language was scanned for template
usages, rather than the whole page. But when the page was saved this
section was saved as the entire page text. Now for simplicity the
entire page is scanned.